If we're unable to create the default work directory, then we're unable to create the StandardTemplateManager, so throw an exception.
git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk/jamon@981 6afd1a24-8710-0410-9bc4-c69fca332c4c

+++ core/src/org/jamon/StandardTemplateManager.java
@@ -373,7 +373,11 @@ public class StandardTemplateManager
                      "jamon"
                      + (new java.util.Random().nextInt(100000000))
                      + ".tmp");
-        workDir.mkdirs();
+        if (! workDir.mkdirs())
+        {
+            throw new JamonException("Unable to create default work directory "
+                                     + workDir);
+        }
         workDir.deleteOnExit();
         return workDir.getCanonicalPath();
     }
